Render map animated objects and fix weapon bugs
- Load the required files for the animated map objects.
- Calculate team goals based on powerball arena settings.
- Fix weapon spawning so it uses pristine heading from the packet
  instead of the player's heading that could have been ignored in the
  packet containing the weapon.
- Fix weapons being timed out when the owner dies.
- Fix non-bouncing bullets so they explode when hitting walls.
- Fix shrapnel so the non-bouncing ones explode when hitting walls.
- Fix shrapnel so their spawn timer is based on the bomb's update tick
  instead of the global tick. This slightly improves the simulation in
  cases where the bomb is being updated multiple times in one tick.
- Fix Repel animation so it doesn't sometimes loop.

+++ src/render/text_renderer.rs
@@ -120,8 +120,8 @@ impl TextRenderer {
                     mvp,
                     camera.scale,
                     renderable,
-                    current_x,
-                    current_y,
+                    current_x as f32,
+                    current_y as f32,
                     layer,
                 );
